LET-10 Add MainFragment Sample code and layout
- 메인 화면에 사용 될 레이아웃 구현
- 컨트리뷰션/뱃지는 기본 틀만 잡아둠
- ProgressBar와 관련된 부분은 대략적인 샘플코드 첨부함
