--- app/src/main/java/com/github/caterpie/letsgitit/MainActivity.kt
@@ -1,14 +1,0 @@
-package com.github.caterpie.letsgitit
-
-import android.os.Bundle
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.github.caterpie.letsgitit.base.BaseActivity
-import com.github.caterpie.letsgitit.databinding.ActivityMainBinding
-
-class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::inflate) {
-	override fun onCreate(savedInstanceState: Bundle?) {
-		val splashScreen = installSplashScreen()
-		super.onCreate(savedInstanceState)
-
-	}
-}
